imported header files for web scraper; added checking for only 1 decimal point and 2 decimal places for text fields; added "Start Over" button

+++ GasShare/RouteSearchViewController.swift
@@ -20,7 +20,7 @@ class RouteSearchViewController: UIViewController {
     @IBOutlet weak var endSearchBar: UISearchBar!
     @IBOutlet weak var containerView: UIView!
     var routeMapViewController: RouteMapViewController!
-    var gasMileage: Int!
+    var gasMileage: Double!
     var gasPrice: Double!
     var routeDistance: Double = -1
